restructure application root component
This adds a new component per main view: Editor (form), Data, and Viewer
and factors out code from the root application component to the relevant
new components. The root component is now basically just for navigation,
using tabs. The existing form editing and form data display functionality
have moved to the MainForm component (aka Editor display), the MainData
component currently contains minimal functionality but will allow fetching
of graph data and loading it into the internal representation to be displayed
and selected from in editors, and the MainViewer component will eventually
contain functionality to render data.
