+++ src/components/MainViewer.vue
@@ -1,0 +1,21 @@
+<template>
+    <v-container fluid>
+        <div style="text-align: center;">
+            <h2>Under construction...</h2>
+            <br>
+            <v-img
+            class="mx-auto"
+            width="40%"
+            src="../assets/undraw_under_construction.svg"
+            ></v-img>
+        </div>
+  </v-container>
+</template>
+
+
+<script setup>
+  import { ref, onMounted, onBeforeMount, provide } from 'vue'
+</script>
+
+
+
